requirements.txt
psutil
pandas
Pillow
requests
beautifulsoup4